krillin: add a script that fixes the build jack compilation error
Since in android nougat google is using the jack compilation method, the old computers may have problems with it (at least my pc)
Already added the build_with_jack to false, but it wasn't fixing the compilation error.
The fix was kill the jack server and reestart it.
So I created a .sh script that does that for you! All you need to know for start the script is in Readme